Restructure repository with git mv
