--- Makefile
@@ -1,14 +1,0 @@
-USER_LIB_PATH=./libraries
-
-TARGET = ayab
-ifeq "$(MACHINETYPE)" "KH910"
-CPPFLAGS += -DKH910
-else
-ifeq "$(MACHINETYPE)" "KH930"
-CPPFLAGS += -DKH930
-else
-$(info "Available values for MACHINETYPE: KH910 KH930")
-endif
-endif
-
-include $(ARDMK_DIR)/Arduino.mk
