+++ src/ayab/knitter.cpp
