+++ src/ayab/knitter.h
@@ -24,7 +24,7 @@ This file is part of AYAB.
 
 #include <Arduino.h>
 
-#include "libraries/PacketSerial/src/PacketSerial.h"
+#include "../../libraries/PacketSerial/src/PacketSerial.h"
 
 #include "beeper.h"
 #include "debug.h"
