+++ src/ayab/serial_encoding.cpp
