+++ src/ayab/serial_encoding.h
