+++ src/ayab/solenoids.cpp
