+++ src/ayab/solenoids.h
