+++ test/mocks/knitter/beeper_mock.cpp
@@ -1,5 +1,5 @@
 #include "beeper_mock.h"
-#include "../../../beeper.h"
+#include "beeper.h"
 
 static BeeperMock *gBeeperMock = NULL;
 BeeperMock *beeperMockInstance() {
