+++ test/test_beeper.cpp
@@ -1,6 +1,6 @@
 #include "gtest/gtest.h"
 
-#include "../beeper.h"
+#include "beeper.h"
 
 using ::testing::Return;
