+++ test/test_encoders.cpp
@@ -1,6 +1,6 @@
 #include "gtest/gtest.h"
 
-#include "../encoders.h"
+#include "encoders.h"
 
 using ::testing::Return;
