+++ test/test_serial_encoding.cpp
@@ -1,6 +1,6 @@
 #include "gtest/gtest.h"
 
-#include "../serial_encoding.h"
+#include "serial_encoding.h"
 #include "serial_encoding/knitter_mock.h"
 
 using ::testing::Return;
